Feat(FirstClient) create shell library with layout
Generate shell library for “FirstClient” application. Main purpose of these types of shell libraries (also application shell libraries) is to define layout of part of application (or, such as in this case, of whole application) and setup routes for that part.

For instance, more application shells could be used when application is divided into 2 parts. First one is available for all visitors (and includes eg. login form). Second part has totally different layout and is available only to logged users.