feat(FirstClient:Products) create shell library
Generate shell library for products segment. Main purpose of these shell libraries is to define routes and pages for one segment (usually but definitely not necessarily pages to handle single CRUD api resource).

This libraries should export only one module, which is meant to be lazily imported in application (or application shell library) router module.

Also notice that library (shell) router module has no direct routes itself. Instead of that, all pages (routes) are defined in own router modules. The only purpose of this is preparation to easier migration to lazy load all pages. Unfortunately this is not supported by Angular yet, so pages must be imported directly (and lazy loaded is only this “bundle” of pages).